feat: remove hashing check
The hashing check is flaky, it doesn't work when you are using the
checkout@v6 action because some kind of git credentials file.

Ideally the copy function is changed to copy over all necessary files,
but we can't spend more time on this, hence we will just get rid of the
hashing check.

Refs: PNT-1054